+++ src/manager/schemas/admin/infer-contact/post.js
@@ -1,0 +1,12 @@
+module.exports = () => ({
+  email: {
+    types: ['string'],
+    default: '',
+    required: false,
+  },
+  emails: {
+    types: ['array'],
+    default: [],
+    required: false,
+  },
+});